fixed build path
