+++ package.json
@@ -1,0 +1,43 @@
+{
+    "name": "sessiontimeoutsave",
+    "widgetName": "SessionTimeOutSave",
+    "version": "1.0.0",
+    "description": "Session Timeout Save",
+    "copyright": "",
+    "author": "",
+    "engines": {
+        "node": ">=16"
+    },
+    "license": "Apache-2.0",
+    "config": {
+        "projectPath": "../../",
+        "mendixHost": "http://localhost:8081",
+        "developmentPort": 3000
+    },
+    "packagePath": "avertra",
+    "scripts": {
+        "start": "pluggable-widgets-tools start:server",
+        "dev": "pluggable-widgets-tools start:web",
+        "build": "pluggable-widgets-tools build:web",
+        "lint": "pluggable-widgets-tools lint",
+        "lint:fix": "pluggable-widgets-tools lint:fix",
+        "prerelease": "npm run lint",
+        "release": "pluggable-widgets-tools release:web"
+    },
+    "devDependencies": {
+        "@mendix/pluggable-widgets-tools": "^9.24.0",
+        "@types/big.js": "^6.0.2"
+    },
+    "dependencies": {
+        "sweetalert2": "^11.7.12",
+        "universal-cookie": "^4.0.4"
+    },
+    "resolutions": {
+        "react": "18.2.0",
+        "react-native": "0.70.7"
+    },
+    "overrides": {
+        "react": "18.2.0",
+        "react-native": "0.70.7"
+    }
+}
